Wait for child stream close events
Previously we waited just for the child process closed event,
which was documented to be emitted when the streams where closed.
With this commit, we wait for all of events.

This is whistling in the dark, hoping for some magic that will fix
the occasional truncated logs problem.

This commit also includes some instrumentation to show the order
that these events are received. The typical order seems to be:

1. "child stderr close"
2. "child stdout close"
3. "child process exit"
4. "child process close"

However, sometimes the order is:

1. "child stderr close"
2. "child process exit"
3. "child process close"
4. "child stdout close"

If log truncation happened only with child stdout, I might rationalize
the this ordering of events could be the smoking gun for the bug. But we have
been observing log truncation when cranking up the `debug` log output, which
has been using `stderr`.